moved pkg folder to the root
